+++ websites/site-template/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @commercetools-website/site-template
 
+## 5.0.2
+
+### Patch Changes
+
+- Updated dependencies [[`a540653f`](https://github.com/commercetools/commercetools-docs-kit/commit/a540653f85e3b396a1b5a937997bd54aaa8fb772), [`c3e43bb4`](https://github.com/commercetools/commercetools-docs-kit/commit/c3e43bb46941ebc4805e0a70f5043a34346b7053)]:
+  - @commercetools-docs/gatsby-theme-docs@18.1.0
+
 ## 5.0.1
 
 ### Patch Changes